+++ crates/orchestrator/src/jobs/da_job/mod.rs
@@ -62,6 +62,7 @@ pub struct DaJob;
 
 #[async_trait]
 impl Job for DaJob {
+    #[tracing::instrument(fields(category = "da"), skip(self, _config, metadata))]
     async fn create_job(
         &self,
         _config: Arc<Config>,
@@ -81,6 +82,7 @@ impl Job for DaJob {
         })
     }
 
+    #[tracing::instrument(fields(category = "da"), skip(self, config))]
     async fn process_job(&self, config: Arc<Config>, job: &mut JobItem) -> Result<String, JobError> {
         let block_no = job
             .internal_id
@@ -145,6 +147,7 @@ impl Job for DaJob {
         Ok(external_id)
     }
 
+    #[tracing::instrument(fields(category = "da"), skip(self, config))]
     async fn verify_job(&self, config: Arc<Config>, job: &mut JobItem) -> Result<JobVerificationStatus, JobError> {
         Ok(config
             .da_client()
@@ -167,6 +170,7 @@ impl Job for DaJob {
     }
 }
 
+#[tracing::instrument(skip(elements))]
 pub fn fft_transformation(elements: Vec<BigUint>) -> Vec<BigUint> {
     let xs: Vec<BigUint> = (0..*BLOB_LEN)
         .map(|i| {