+++ crates/orchestrator/src/jobs/state_update_job/mod.rs
@@ -72,6 +72,7 @@ pub enum StateUpdateError {
 pub struct StateUpdateJob;
 #[async_trait]
 impl Job for StateUpdateJob {
+    #[tracing::instrument(fields(category = "state_update"), skip(self, _config, metadata))]
     async fn create_job(
         &self,
         _config: Arc<Config>,
@@ -99,6 +100,7 @@ impl Job for StateUpdateJob {
         })
     }
 
+    #[tracing::instrument(fields(category = "state_update"), skip(self, config))]
     async fn process_job(&self, config: Arc<Config>, job: &mut JobItem) -> Result<String, JobError> {
         let attempt_no = job
             .metadata
@@ -149,6 +151,7 @@ impl Job for StateUpdateJob {
     /// Status will be verified if:
     /// 1. the last settlement tx hash is successful,
     /// 2. the expected last settled block from our configuration is indeed the one found in the provider.
+    #[tracing::instrument(fields(category = "state_update"), skip(self, config))]
     async fn verify_job(&self, config: Arc<Config>, job: &mut JobItem) -> Result<JobVerificationStatus, JobError> {
         let attempt_no = job
             .metadata