+++ crates/orchestrator/src/main.rs
@@ -2,13 +2,18 @@ use dotenvy::dotenv;
 use orchestrator::config::init_config;
 use orchestrator::queue::init_consumers;
 use orchestrator::routes::app_router;
+use orchestrator::telemetry::{setup_analytics, shutdown_analytics};
+
 use utils::env_utils::get_env_var_or_default;
 
 /// Start the server
 #[tokio::main]
 async fn main() {
     dotenv().ok();
-    tracing_subscriber::fmt().with_max_level(tracing::Level::INFO).with_target(false).init();
+
+    // Analytics Setup
+
+    let meter_provider = setup_analytics();
 
     // initial config setup
     let config = init_config().await;
@@ -24,4 +29,7 @@ async fn main() {
 
     tracing::info!("Listening on http://{}", address);
     axum::serve(listener, app).await.expect("Failed to start axum server");
+
+    // Analytics Shutdown
+    shutdown_analytics(meter_provider);
 }