+++ crates/utils/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod collections;
 pub mod env_utils;
+pub mod metrics;
 pub mod settings;
 
 /// Evaluate `$x:expr` and if not true return `Err($y:expr)`.