+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Added
 
+- Telemetry tracing and metrics.
 - e2e flow test
 - database timestamps
 - alerts module.